pyproject.toml
[tool.poetry]
name = "peakrdl_markdown"
version = "1.1.0"
description = "Export Markdown description from the systemrdl-compiler register model"
authors = ["Marek Pikuła <marek@serenitycode.dev>"]
maintainers = ["Marek Pikuła <marek@serenitycode.dev>"]
license = "GPL-3.0-only"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Electronic Design Automation (EDA)",
]
repository = "https://github.com/SystemRDL/PeakRDL-Markdown"
documentation = "https://peakrdl-markdown.readthedocs.io"
readme = "README.md"
packages = [{ include = "peakrdl_markdown", from = "src" }]

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"

[tool.poetry.dependencies]
python = "^3.7.0"
systemrdl-compiler = "^1.30.1"
peakrdl-cli = "^1.5.0"
py-markdown-table = [
    { version = "<1.0.0", python = "<3.8" },
    { version = "^1.3.0", python = ">=3.8" },
]

[tool.poetry.group.dev.dependencies]
pre-commit = { version = "^4.4.0", python = ">=3.12" }

[tool.poetry.group.lint.dependencies]
black = { version = "^25.11.0", python = ">=3.12" }
isort = { version = "^7.0.0", python = ">=3.12" }
mypy = { version = "^1.18.2", python = ">=3.12" }
pydocstyle = { version = "^6.3.0", python = ">=3.12" }
ruff = { version = "^0.14.4", python = ">=3.12" }

[tool.poetry.group.test.dependencies]
pytest = [
    { version = "<8.1.1", python = "<3.8" },
    { version = "^8.3.0", python = ">=3.8" },
]
pytest-cov = [
    { version = "<5.0.0", python = "<3.8" },
    { version = "^5.0.0", python = ">=3.8" },
]
coverage = "^6.3"
coveralls = "^3.3.1"

[tool.poetry.group.doc.dependencies]
linuxdoc = { version = "^20240924", python = ">=3.12" }
m2r2 = { version = "^0.3.4", python = ">=3.12" }
pygments-systemrdl = { version = "^1.3.0", python = ">=3.12" }
Sphinx = { version = "^8.2.3", python = ">=3.12" }
sphinx-rtd-theme = { version = "^3.0.2", python = ">=3.12" }

[tool.poetry.plugins."peakrdl.exporters"]
"markdown" = "peakrdl_markdown.__peakrdl__:Exporter"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.pydocstyle]
convention = "pep257"
